Add the gt GFF validation module - run on the summary gff
Run the genometools GFF validation

If there are any errors with the GFF, these will be added to a log file for further inspection, but an invalid GFF won't make the pipeline fail.
Doing so would be a waste of compute and not something we could take care of at this point of the execution.
This is based on the assumption that we will test the pipeline with loads of different kinds of assemblies and by then
we should have ironed out any issues.

The log file will be used in the automation system to flag the jobs for inspection.

But, I'm up to discuss this